+++ test/integration/configuration/aws_inspec_config.rb
@@ -209,7 +209,36 @@ module AWSInspecConfig
       # Only test execution is affected by this flag, resource creation via terraform is unaffected.
       # Default behaviour is for this to be disabled, enable by changing the below flag.
       aws_enable_cli_calls: 0,
-      aws_route_53_zone: "aws-route53-zone-#{add_random_string}"
+      aws_route_53_zone: "aws-route53-zone-#{add_random_string}",
+      tgw_route_cidr_block: '0.0.0.0/15',
+      tgw_route_cidr_block_blockhole: '0.0.0.0/16',
+      aws_route_53_zone: "aws-route53-zone-#{add_random_string}",
+      aws_launch_template_name: "launch_template-#{add_random_string}",
+      aws_launch_template_core: 4,
+      aws_launch_template_threads_per_core: 2,
+      aws_launch_template_cpu_credits: "standard",
+      aws_launch_template_volume_size: 20,
+      aws_launch_template_instance_profile: "test-profile",
+      aws_launch_template_resource_type: "instance",
+      aws_launch_template_tag_name: "test",
+      aws_launch_template_instance_type: "t2.micro",
+      aws_launch_template_kernel_id: "test_kernel_id",
+      aws_launch_template_key_name: "test_key_name",
+      aws_vpn_gw_name: 'inspec_vpn_gw',
+      aws_db_parameter_group_name: 'inspec-mysql-db-group-param',
+      aws_db_parameter_group_family_name: 'mysql5.6',
+      aws_db_parameter_group_description: 'Inspec Managed parameter group for mysql5.6',
+      aws_db_option_group_name: 'option-group-test-terraform2',
+      aws_db_option_group_description: 'Terraform Option Group',
+      aws_db_option_group_engine_name: 'sqlserver-ee',
+      aws_redshift_cluster_identifier: "redshift-cluster-#{add_random_string}",
+      aws_redshift_parameter_group_name: "group#{add_random_string}",
+      aws_redshift_parameter_group_family: "redshift-1.0",
+      aws_vpn_gw_name: 'inspec_vpn_gw',
+      aws_network_acl_cidr_block: '10.3.0.0/18',
+      aws_network_acl_name: 'inspec-nw-acl',
+      acl_egress_rule_number: 100,
+      acl_ingress_rule_number: 100,
   }
 
   def self.config